move doc button to resource panel
