+++ apps/src/music/views/MusicLabView.tsx
@@ -303,6 +303,8 @@ const MusicLabView: React.FunctionComponent<MusicLabViewProps> = ({
     return <MusicPlayView setPlaying={setPlaying} />;
   }
 
+  const docUrl = Blockly.showBlockHelp ? '/docs/ide/music' : undefined;
+
   return (
     <div
       id="music-lab"
@@ -354,6 +356,7 @@ const MusicLabView: React.FunctionComponent<MusicLabViewProps> = ({
             hideContinueIfDisabled={true}
             hideNavigation={false}
             styleNavigationAsBubble={true}
+            documentationUrl={docUrl}
           />
         </div>
